+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- Configurable span batching [#5](https://github.com/JohnDoneth/spandex_otlp/pull/5).
 - Handle conversion of Spandex error fields [#4](https://github.com/JohnDoneth/spandex_otlp/pull/4).
 
 ### v0.1.0-rc.1